requirements.txt
fastapi
uvicorn
pydantic
python-multipart
torch
torchvision
scikit-learn
pandas
numpy
pillow
pytest
requests